requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
waitress==2.1.2
keras==3.4.1
tensorflow==2.16.2
numpy==1.23.5
pillow==9.3.0
scikit-learn==1.2.2
